move all detour to harmony
